# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Los Jurones Domino Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Matches ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, name, wins, losses, totalPoints, winRate, currentStreak, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        wins:
          type: integer
        losses:
          type: integer
        totalPoints:
          type: integer
        winRate:
          type: number
        currentStreak:
          type: integer
        createdAt:
          type: string
    PlayerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        avatarUrl:
          type: string
    PlayerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        avatarUrl:
          type: ["string", "null"]
        wins:
          type: integer
        losses:
          type: integer
        totalPoints:
          type: integer
        currentStreak:
          type: integer

    # ─── Match ────────────────────────────────────────────────────────────
...