# lib/api-zod/src/generated/types/playerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Los Jurones Domino Club API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerInput {
  /** @minLength 1 */
  name: string;
  avatarUrl?: string;
}