# lib/api-zod/src/generated/types/playerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Los Jurones Domino Club API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  wins?: number;
  losses?: number;
  totalPoints?: number;
  currentStreak?: number;
}